Make DiffView and FieldView subclassable
